code/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
ast-tools==0.1.8
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
bids-validator==1.14.7.post0
bidsschematools==1.0.5
bitarray==3.3.2
bottle==0.13.3
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.1.8
cytoolz==1.0.1
decorator==5.2.1
docopt==0.6.2
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
executing==2.2.0
formulaic==1.1.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.1
hexbytes==1.3.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
interface-meta==1.3.0
ipdb==0.13.13
ipython==8.12.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
libcst==1.7.0
lxml==5.4.0
matplotlib-inline==0.1.7
multidict==6.4.3
networkx==3.4.2
nibabel==5.3.2
nilearn==0.11.1
num2words==0.5.14
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pybids==0.19.0
pycryptodome==3.22.0
pycups==2.0.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.40
stack-data==0.6.3
threadpoolctl==3.6.0
tifffile==2025.3.30
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
universal_pathlib==0.2.6
urllib3==2.4.0
wcwidth==0.2.13
web3 @ git+https://github.com/ethereum/web3.py.git@a08787d08a8f30a099ea1daf1a5e6575dff07681
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
